refactor(auth): isolate school redirect
